Clarify Incident Priority
Just to clarify the priority 3 in comparison to 2, as the first line is similar. Also, because in the portal list of incident priorities, there is only one option for priority 3 which is "isolated to a single user". Therefore I have added the words: "for a single user" so as to clarify to customers that it is a priority for a single user and priority 2 is for multiple users