Update enbl-sco-mtf-nas.md
I have merged the NAS and HDD service scopes into one as it was confusing to have two when the majority of the information was the same. I have also included USBs as a device, and confirmed the cost rather than referring the customer to the SFIA rate card which seems a bit noddy.