Avoid writing credentials into shell commands
It is safer to use `read -s` to put passwords into environment variables
because:
1. People looking over your shoulder can't read them off your screen.
2. They don't appear in your shell history.

It's worth noting it is still not totally secure.
If you don't run the `unset` command, people using your shell,
malicious shell commands and shell commands that don't realise the
sensitivity of the environment variables can read the password.