Prefer https to http
We should always prefer https links to http. Especially when we link to
downloads.

This is to prevent man in the middle attacks. This website uses https,
so as long as warnings haven't been clicked through it is secure. But if
we link users off to an http website, an attacker can then start viewing
and controlling what our users see.

I've not changed all links.
* I've left the index.json because it can be regenerated automatically.
* I've left some http links because either they worked http but not
https, or because they don't work either way. We should set up a scanner
to verify our links are all live.