Rename pngs to lower case extension
Having differences in the file-names referenced in the markdown, and
the file-names in the repo generates warnings when building on case
sensitive file systems (such as Linux, especially github actions).

By having the png file extension consistently lower case, we can prevent
the warnings and move a step closer to having github-actions build the
docs for us automatically (continuous delivery).