Build the tocs using github actions
I had to make some changes to the createTOC powershell file.

First I had to change the path joining to be platform agnostic so that
it would work on Linux.

Second I had to change the output to JSON.
All JSON is valid YAML, so this isn't actually a drastic change.
For some reason, when outputting as YAML, the headings were all off by
one.