Link to Portal OpenAPI reference
The OpenAPI reference is currently in development, but once complete, it is likely to be more useful to users.

It allows clients to be generated.
It's neater, the responses are hidden until you click through to them.
It's more consistent - curl examples are constructed automatically and correctly. 

---

Open question: Should the json file be moved to this repository?