Remove Portal /api/my_vms endpoint documentation
Although not in the documentation, when you call it it returns

```json
"warning":"This endpoint has been deprecated. Please use https://portal.skyscapecloud.com/api/accounts/:account_id/compute_service in future"
```

And has been since Portal 1.37.1 (Wed Oct 26 2016).

At the very least, it shouldn't be the second in the list and not say it's deprecated.