Update oshift-how-expand-persistent-vols.md
Made a minor change to call out the differences in v4. 

However, I didn't want to fully document it for v4 - At this time, the only way to scale the pvc for logging in v4 (without deleting data) is to manually scale the PVC. However, in future v4 versions, I'd expect the openshift-logging operator to be able to handle expansion of the pvc?