Remove references to Data Transfer Facility
The Data Transfer Facility has been retired. This pull requests removes references to the DTF from the Knowledge Centre.
I've kept a single reference in the Glossary that points customers to the Mass Transfer Facility just in case anyone searches for DTF. We can remove this the next time the Glossary is updated.